requirements.txt
fastapi
httpx
pydantic
pydantic[email]
pytest
python-dotenv
slowapi
sqlmodel
streamlit
streamlit_pydantic
uvicorn